requirements.txt
python-dotenv
neo4j
tabulate
flask
flask-cors
